Added compose example and better readmes
