+++ HelloAgainOptimization/Dockerfile
@@ -10,18 +10,14 @@ RUN dotnet publish \
     -r linux-musl-x64 \
     --self-contained true \
     -p:PublishTrimmed=true \
-    -p:PublishReadyToRun=true
+    -p:PublishSingleFile=true
+    #-p:PublishReadyToRun=true
 
 FROM alpine:3.12 as runtime-machine
 COPY --from=build-machine /app /app
-# As we saw from the previous Dockerfile our container doesn't run! The reason is simple,
-# our Alpine base image is so stripped down it doesn't have the libraries that .NET Core
-# requires. The fix is simple, we just have to install the.
 
 RUN apk update && apk add --no-cache libstdc++ libgcc libintl
 
 ENV DOTNET_SYSTEM_GLOBALIZATION_INVARIANT true
 WORKDIR /app
 ENTRYPOINT ["./HelloWorld"]
-# And the result: A 46MB image. However note that size before we added the required libs for
-# .NET Core, the image was <6MB.
