+++ HelloAgainOptimization/run.ps1
@@ -1,0 +1,1 @@
+docker run -ti helloagainoptimization
