--- HelloCompose/ConsoleUI/IApplication.cs
@@ -1,7 +1,0 @@
-﻿namespace ConsoleUI
-{
-  public interface IApplication
-  {
-    void Run();
-  }
-}
