--- HelloCompose/ConsoleUI/Program.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace ConsoleUI
-{
-  class Program
-  {
-    static void Main(string[] args)
-    {
-      Console.WriteLine("Hello World!");
-    }
-  }
-}
