+++ HelloCompose/DemoApplication/ConsoleUI/Application.cs
@@ -1,0 +1,24 @@
+﻿using DemoLibrary;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace ConsoleUI
+{
+  public class Application : IApplication
+  {
+    IBusinessLogic _businessLogic;
+
+    public Application(IBusinessLogic businessLogic)
+    {
+      _businessLogic = businessLogic;
+    }
+
+    public void Run()
+    {
+      _businessLogic.ProcessData();
+    }
+  }
+}
