+++ HelloCompose/DemoApplication/ConsoleUI/ContainerConfig.cs
@@ -1,0 +1,28 @@
+﻿using Autofac;
+using DemoLibrary;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Reflection;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace ConsoleUI
+{
+  public static class ContainerConfig
+  {
+    public static IContainer Configure()
+    {
+      var builder = new ContainerBuilder();
+
+      builder.RegisterType<Application>().As<IApplication>();
+      builder.RegisterType<BusinessLogic>().As<IBusinessLogic>();
+
+      builder.RegisterAssemblyTypes(Assembly.Load(nameof(DemoLibrary)))
+          .Where(t => t.Namespace.Contains("Utilities"))
+          .As(t => t.GetInterfaces().FirstOrDefault(i => i.Name == "I" + t.Name));
+
+      return builder.Build();
+    }
+  }
+}
