+++ HelloCompose/DemoApplication/ConsoleUI/IApplication.cs
@@ -1,0 +1,7 @@
+﻿namespace ConsoleUI
+{
+  public interface IApplication
+  {
+    void Run();
+  }
+}
