+++ HelloCompose/DemoApplication/DemoLibrary/IBusinessLogic.cs
@@ -1,0 +1,7 @@
+﻿namespace DemoLibrary
+{
+  public interface IBusinessLogic
+  {
+    void ProcessData();
+  }
+}
