+++ HelloCompose/DemoApplication/DemoLibrary/Utilities/ILogger.cs
@@ -1,0 +1,7 @@
+﻿namespace DemoLibrary.Utilities
+{
+  public interface ILogger
+  {
+    void Log(string message);
+  }
+}
