+++ HelloCompose/DemoApplication/DemoLibrary/Utilities/Logger.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace DemoLibrary.Utilities
+{
+  public class Logger : ILogger
+  {
+    public void Log(string message)
+    {
+      Console.WriteLine($"DBG: { message }");
+    }
+  }
+}
