--- HelloCompose/DemoLibrary/IBusinessLogic.cs
@@ -1,7 +1,0 @@
-﻿namespace DemoLibrary
-{
-  public interface IBusinessLogic
-  {
-    void ProcessData();
-  }
-}
