--- HelloCompose/DemoLibrary/Utilities/ILogger.cs
@@ -1,7 +1,0 @@
-﻿namespace DemoLibrary.Utilities
-{
-  public interface ILogger
-  {
-    void Log(string message);
-  }
-}
