--- HelloCompose/DemoLibrary/Utilities/Logger.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace DemoLibrary.Utilities
-{
-  public class Logger : ILogger
-  {
-    public void Log(string message)
-    {
-      Console.WriteLine($"DBG: { message }");
-    }
-  }
-}
