+++ HelloCompose/build.ps1
@@ -1,0 +1,1 @@
+docker-compose build --no-cache
