+++ HelloCompose/run.ps1
@@ -1,0 +1,1 @@
+docker-compose up
