+++ HelloOptimization/build.ps1
@@ -1,0 +1,1 @@
+docker build --tag hellooptimization --file Dockerfile .
